Move towards using a worker
