+++ packages/cspell-eslint-plugin/src/_auto_generated_/options.schema.json
@@ -37,10 +37,6 @@
         {
           "additionalProperties": false,
           "properties": {
-            "addWords": {
-              "description": "**Experimental**: Provide a fix option to add words to the file.\n\nNote: this does not yet work perfectly.",
-              "type": "boolean"
-            },
             "path": {
               "description": "Path to word list file. File format: 1 word per line",
               "type": "string"
