+++ packages/cspell-eslint-plugin/src/spellCheck.ts
